// src/types.ts
export interface Coin {
    availableSupply: number;
    explorers: string[];
    icon: string;
    id: string;
    marketCap: number;
    name: string;
    price: number;
    priceBtc: number;
    priceChange1d: number;
    priceChange1h: number;
    priceChange1w: number;
    rank: number;
    redditUrl: string;
    symbol: string;
    totalSupply: number;
    twitterUrl: string;
    volume: number;
    websiteUrl: string;
  }
  